… bbhash.h