Most gains in superbuckets creation were obtained by ensuring no div/mod instruction were emitted by the compiler where rshifts/maskings is possible (aad3b03)

The new hashes functions needs to be tested on bigger dataset I think (c091c29)

I needed to revert d70cf79 (91e3c18) as something went terribly wrong.

The multi-threading in bbhash is removed in 3f4990c as it is not compatible with many MPHF, and IMO it was coded using unspeakable human body parts.

I managed to get good improvements on multi-threading but I have some doubts about the correctness as I don't know the exact workings of the algo (some things are exclusive in my mind but may not be & the other way around). See the blight.h part of 3f4990c and 4ed7f2f 

Ideally the `read_super_buckets()`'s loop should be parallelized instead of the `create_mphf()` & `fill_positions()` ones. For some reason `str2bool()` is doing non-exlusive access as if a minimizer is shared between superbuckets ?

Overall it went from
```
I use -g /home/piezo/data/cel.unitigs.fa -q /home/piezo/data/cel.unitigs.fa -k 31 -m  9 -n  0 -s  3 -t 4 -b 6
Super bucket created: 28.9634 seconds.
----------------------------------------------------------------
----------------------INDEX RECAP----------------------------
Kmer in graph: 94,006,920
Super Kmer in graph: 8,184,928
Average size of Super Kmer: 11
Total size of the partitionned graph: 339,554,760
Total size of the partitionned graph: 339,554,784
Largest MPHF: 543,527
Largest Bucket: 1,511,971
Size of the partitionned graph (MBytes): 80
Total Positions size (MBytes): 111
Size of the partitionned graph (bit per kmer): 7.22404
Total Positions size (bit per kmer): 9.92692
TOTAL Bits per kmer (without bbhash): 17.151
TOTAL Bits per kmer (with bbhash): 21.151
TOTAL Size estimated (MBytes): 237.028
Indexes created: 226.856 seconds.
The whole indexing took me 255.819 seconds.
-----------------------QUERY RECAP 2----------------------------
Good kmer: 93,995,065
Erroneous kmers: 0
Query performed: 94,006,897
The whole QUERY took me 68.7343 seconds.
I am glad you are here with me. Here at the end of all things.
./bench_blight -g ~/data/cel.unitigs.fa -q ~/data/cel.unitigs.fa -k 31 -t 4  635.37s user 195.82s system 255% cpu 5:24.72 total
```

to 
```
I use -g /home/piezo/data/cel.unitigs.fa -q /home/piezo/data/cel.unitigs.fa -k 31 -m  9 -n  0 -s  3 -t 4 -b 6
Super bucket created: 18.7371 seconds.
----------------------------------------------------------------
----------------------INDEX RECAP----------------------------
Kmer in graph: 94,006,920
Super Kmer in graph: 8,166,653
Average size of Super Kmer: 11
Total size of the partitionned graph: 339,006,510
Total size of the partitionned graph: 339,006,528
Largest MPHF: 1,905,164
Largest Bucket: 4,434,824
Size of the partitionned graph (MBytes): 80
Total Positions size (MBytes): 111
Size of the partitionned graph (bit per kmer): 7.21237
Total Positions size (bit per kmer): 9.97155
TOTAL Bits per kmer (without bbhash): 17.1839
TOTAL Bits per kmer (with bbhash): 21.1839
TOTAL Size estimated (MBytes): 237.398
Indexes created: 25.3978 seconds.
The whole indexing took me 44.1349 seconds.
-----------------------QUERY RECAP 2----------------------------
Good kmer: 93,981,062
Erroneous kmers: 0
Query performed: 94,006,897
The whole QUERY took me 23.331 seconds.
I am glad you are here with me. Here at the end of all things.
./bench_blight -g ~/data/cel.unitigs.fa -q ~/data/cel.unitigs.fa -k 31 -t 4  197.20s user 0.90s system 292% cpu 1:07.62 total
```

Note how the system time is spared from the horrendous `pthread_create()`s that were inflicted by bbhash...